Add files via upload
Here is a basic recommendation system. The dataset contains 198 different books and the ratings given by users.
K- Nearest Neighbors is used to calculate the similarity between books and getting recommendations for the top 5 rated books, taking threshold rating as 5.
Count of users is also calculated to whom these books will be recommended.